Move private class from BaseContext to ContextFactory class instead
