+++ src/Traces.Data/TracesDbContext.cs
@@ -6,11 +6,6 @@ namespace Traces.Data
 {
     public class TracesDbContext : BaseContext
     {
-        public TracesDbContext(DbContextOptions options)
-            : base(options)
-        {
-        }
-
         public TracesDbContext(DbContextOptions options, IRequestContext requestContext)
             : base(options, requestContext)
         {
